+++ arcade/gl/buffer.py
@@ -133,11 +133,14 @@ class Buffer:
         :rtype: bytes
         """
         if size == -1:
-            size = self._size
+            size = self._size - offset
 
         # Catch this before confusing INVALID_OPERATION is raised
         if size < 1:
-            raise ValueError("Attempting to read 0 or less bytes from buffer")
+            raise ValueError(
+                "Attempting to read 0 or less bytes from buffer: "
+                f"buffer size={self._size} | params: size={size}, offset={offset}"
+            )
 
         # Manually detect this so it doesn't raise a confusing INVALID_VALUE error
         if size + offset > self._size: